+++ Backend/bookverse/src/main/java/com/licenta/bookverse/service/TrailsService.java
@@ -28,7 +28,7 @@ public class TrailsService {
     private final BookService bookService;
     private final NotificationService notificationService;
 
-    public TrailDTODetails getTrailById(Integer trailId) {
+    public TrailDTODetails getTrailById(Long trailId) {
         Trail trail = trailRepository.findById(trailId)
                 .orElseThrow(() -> new RuntimeException("Trail not found"));
         List<TrailBookDTODetails> bookDtos =
@@ -109,8 +109,6 @@ public class TrailsService {
                         .trailId(trail.getId())
                         .title(trail.getTitle())
                         .description(trail.getDescription())
-                       // .genre(trail.getGenres())
-                      //  .trailBookList(trail.getTrailBooks())
                         .numberOfReadings(trail.getNumberOfReadings())
                         .creatorId(trail.getCreator().getId())
                         .personName(trail.getCreator().getFullName())
@@ -177,7 +175,7 @@ public class TrailsService {
         trailRepository.save(trail);
     }
 
-    public Long updateTrail(Integer trailId, TrailDTO trailDTO) {
+    public Long updateTrail(Long trailId, TrailDTO trailDTO) {
         Trail trail = trailRepository.findById(trailId)
                 .orElseThrow(() -> new RuntimeException("Trail not found"));
